Move NickManager to global scope
