+++ src/main/java/rip/bolt/ingame/managers/NickManager.java
@@ -1,41 +1,35 @@
-package rip.bolt.ingame.ranked;
-
-import static net.kyori.adventure.text.Component.text;
+package rip.bolt.ingame.managers;
 
 import dev.pgm.community.Community;
 import dev.pgm.community.nick.feature.NickFeature;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
-import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.Bukkit;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import rip.bolt.ingame.Ingame;
-import rip.bolt.ingame.ranked.forfeit.PlayerWatcher;
 import rip.bolt.ingame.utils.Messages;
 import tc.oc.pgm.api.event.NameDecorationChangeEvent;
 import tc.oc.pgm.api.integration.Integration;
 import tc.oc.pgm.api.match.MatchScope;
+import tc.oc.pgm.api.party.Competitor;
 import tc.oc.pgm.api.player.MatchPlayer;
 import tc.oc.pgm.events.PlayerJoinMatchEvent;
 
 public class NickManager implements Listener {
 
-  private final PlayerWatcher watcher;
-
-  public NickManager(PlayerWatcher playerWatcher) {
-    this.watcher = playerWatcher;
-  }
+  public NickManager() {}
 
   @EventHandler(priority = EventPriority.NORMAL)
   public void onPlayerJoinMatch(PlayerJoinMatchEvent event) {
     MatchPlayer player = event.getPlayer();
 
-    UUID playerId = player.getId();
-    if (!watcher.isPlaying(playerId)) return;
+    if (!(event.getNewParty() instanceof Competitor)) return;
     if (Ingame.get().getMatchManager().getMatch() == null) return;
     if (Integration.getNick(player.getBukkit()) == null) return;
+
+    UUID playerId = player.getId();
 
     NickFeature nick = Community.get().getFeatures().getNick();
 
@@ -47,10 +41,6 @@ public class NickManager implements Listener {
     event
         .getMatch()
         .getExecutor(MatchScope.LOADED)
-        .schedule(
-            () -> player.sendMessage(Messages.withSeparators(text(
-                "Your nick was removed as you are playing in this match.", NamedTextColor.RED))),
-            2,
-            TimeUnit.SECONDS);
+        .schedule(() -> player.sendMessage(Messages.nickRemoved()), 2, TimeUnit.SECONDS);
   }
 }
