+++ src/main/java/rip/bolt/ingame/ranked/RankedManager.java
@@ -28,7 +28,6 @@ public class RankedManager extends GameManager {
   private final RankManager rankManager;
   private final SpectatorManager spectatorManager;
   private final RequeueManager requeueManager;
-  private final NickManager nickManager;
 
   public RankedManager(MatchManager matchManager) {
     super(matchManager);
@@ -37,7 +36,6 @@ public class RankedManager extends GameManager {
     this.rankManager = new RankManager(matchManager);
     this.spectatorManager = new SpectatorManager(playerWatcher);
     this.requeueManager = new RequeueManager();
-    this.nickManager = new NickManager(playerWatcher);
   }
 
   @Override
@@ -48,15 +46,16 @@ public class RankedManager extends GameManager {
     Bukkit.getPluginManager().registerEvents(this.rankManager, Ingame.get());
     Bukkit.getPluginManager().registerEvents(this.spectatorManager, Ingame.get());
     Bukkit.getPluginManager().registerEvents(this.requeueManager, Ingame.get());
-    Bukkit.getPluginManager().registerEvents(this.nickManager, Ingame.get());
   }
 
   @Override
   public void setup(BoltMatch match) {
     super.setup(match);
     EventsPlugin.get().getTeamManager().clear();
-    for (TournamentTeam team : match.getTeams())
+    for (TournamentTeam team : match.getTeams()) {
       EventsPlugin.get().getTeamManager().addTeam(team);
+    }
+
     playerWatcher.addPlayers(match.getTeams().stream()
         .flatMap(team -> team.getPlayers().stream())
         .map(TournamentPlayer::getUUID)
@@ -70,7 +69,6 @@ public class RankedManager extends GameManager {
     HandlerList.unregisterAll(this.rankManager);
     HandlerList.unregisterAll(this.spectatorManager);
     HandlerList.unregisterAll(this.requeueManager);
-    HandlerList.unregisterAll(this.nickManager);
   }
 
   public PlayerWatcher getPlayerWatcher() {
